+++ spec/requests/api/v1/points_spec.rb
@@ -4,7 +4,7 @@ require 'rails_helper'
 
 RSpec.describe 'Api::V1::Points', type: :request do
   let!(:user) { create(:user) }
-  let!(:points) { create_list(:point, 150, user:) }
+  let!(:points) { create_list(:point, 15, user:) }
 
   describe 'GET /index' do
     context 'when regular version of points is requested' do
@@ -21,7 +21,7 @@ RSpec.describe 'Api::V1::Points', type: :request do
 
         json_response = JSON.parse(response.body)
 
-        expect(json_response.size).to eq(100)
+        expect(json_response.size).to eq(15)
       end
 
       it 'returns a list of points with pagination' do
@@ -31,7 +31,7 @@ RSpec.describe 'Api::V1::Points', type: :request do
 
         json_response = JSON.parse(response.body)
 
-        expect(json_response.size).to eq(10)
+        expect(json_response.size).to eq(5)
       end
 
       it 'returns a list of points with pagination headers' do
@@ -40,7 +40,7 @@ RSpec.describe 'Api::V1::Points', type: :request do
         expect(response).to have_http_status(:ok)
 
         expect(response.headers['X-Current-Page']).to eq('2')
-        expect(response.headers['X-Total-Pages']).to eq('15')
+        expect(response.headers['X-Total-Pages']).to eq('2')
       end
     end
 
@@ -58,7 +58,7 @@ RSpec.describe 'Api::V1::Points', type: :request do
 
         json_response = JSON.parse(response.body)
 
-        expect(json_response.size).to eq(100)
+        expect(json_response.size).to eq(15)
       end
 
       it 'returns a list of points with pagination' do
@@ -68,7 +68,7 @@ RSpec.describe 'Api::V1::Points', type: :request do
 
         json_response = JSON.parse(response.body)
 
-        expect(json_response.size).to eq(10)
+        expect(json_response.size).to eq(5)
       end
 
       it 'returns a list of points with pagination headers' do
@@ -77,7 +77,7 @@ RSpec.describe 'Api::V1::Points', type: :request do
         expect(response).to have_http_status(:ok)
 
         expect(response.headers['X-Current-Page']).to eq('2')
-        expect(response.headers['X-Total-Pages']).to eq('15')
+        expect(response.headers['X-Total-Pages']).to eq('2')
       end
 
       it 'returns a list of points with slim attributes' do
@@ -113,6 +113,20 @@ RSpec.describe 'Api::V1::Points', type: :request do
 
         expect(json_response.first['timestamp']).to be > json_response.last['timestamp']
       end
+    end
+  end
+
+  describe 'PATCH /update' do
+    let!(:point) { create(:point, latitude: 1.0, longitude: 1.0, user:) }
+
+    it 'updates a point' do
+      patch "/api/v1/points/#{point.id}?api_key=#{user.api_key}",
+            params: { point: { latitude: 1.1, longitude: 1.1 } }
+
+      expect(response).to have_http_status(:ok)
+
+      expect(point.reload.latitude).to eq(1.1)
+      expect(point.reload.longitude).to eq(1.1)
     end
   end
 end